refactor(bulletin): 提取 ALBulletinFetchWorker 到独立文件
